pyproject.toml
[tool.poetry]
name = "stock-price-estimator"
version = "00.0.1"
description = ""
authors = ["Alexey Zauzin <lehadnk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
pydantic = "^2.10.6"
tinkoff-investments = "^0.2.0b109"
psycopg2 = "^2.9.10"
clickhouse-driver = "^0.2.9"
scikit-learn = "^1.6.1"
numpy = "^2.2.3"
pandas = "^2.2.3"
clickhouse-connect = "^0.8.15"
pyjwt = "^2.10.1"
xgboost = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
faker = "^36.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
